+++ fastadapter-viewbinding/src/main/kotlin/ca/allanwang/fastadapter/viewbinding/BindingItem.kt
@@ -13,15 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package ca.allanwang.fastadapter.databinding
+package ca.allanwang.fastadapter.viewbinding
 
 import android.content.Context
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.databinding.DataBindingUtil
-import androidx.databinding.ViewDataBinding
+import android.widget.ImageView
+import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
+import androidx.viewbinding.ViewBinding
+import ca.allanwang.kau.fastadapter.viewbinding.R
 import ca.allanwang.kau.logging.KL
 import com.mikepenz.fastadapter.FastAdapter
 import com.mikepenz.fastadapter.GenericItem
@@ -32,44 +34,54 @@ interface VhModel {
     fun vh(): GenericItem
 }
 
-abstract class BindingItem<Binding : ViewDataBinding>(open val data: Any?) :
-    AbstractItem<BindingItem.ViewHolder>(),
-    BindingLayout<Binding> {
+/**
+ * Layout container. Should be implemented in a [BindingItem] companion.
+ */
+interface BindingLayout<Binding : ViewBinding> {
+    val layoutRes: Int
+}
+
+abstract class BindingItem<Binding : ViewBinding>(open val data: Any?) :
+        AbstractItem<BindingItem.ViewHolder>(),
+        BindingLayout<Binding> {
 
     override val type: Int
         get() = layoutRes
 
+    abstract fun createBinding(layoutInflater: LayoutInflater, parent: ViewGroup?): Binding
+
     override fun createView(ctx: Context, parent: ViewGroup?): View {
-        val binding: ViewDataBinding = DataBindingUtil.inflate(
-            LayoutInflater.from(ctx),
-            layoutRes, parent,
-            false
-        )
+        val binding = createBinding(LayoutInflater.from(ctx), parent)
+        setBinding(binding.root, binding)
         return binding.root
     }
 
-    fun getBinding(holder: ViewHolder): Binding? =
-        DataBindingUtil.getBinding<Binding>(holder.itemView)
-
     final override fun bindView(holder: ViewHolder, payloads: MutableList<Any>) {
         super.bindView(holder, payloads)
-        val binding = getBinding(holder) ?: return
+        val binding = holder.getBinding<Binding>()
         binding.bindView(holder, payloads)
-        binding.executePendingBindings()
     }
 
     abstract fun Binding.bindView(holder: ViewHolder, payloads: MutableList<Any>)
 
+    protected fun unbind(vararg textViews: TextView) {
+        textViews.forEach { it.text = null }
+    }
+
+    protected fun unbind(vararg imageViews: ImageView) {
+        imageViews.forEach { it.setImageDrawable(null) }
+    }
+
     final override fun unbindView(holder: ViewHolder) {
         super.unbindView(holder)
-        val binding = DataBindingUtil.getBinding<Binding>(holder.itemView) ?: return
+        val binding = holder.getBinding<Binding>()
         binding.unbindView(holder)
-        binding.unbind()
     }
 
     open fun Binding.unbindView(holder: ViewHolder) {}
 
-    final override fun getViewHolder(v: View): ViewHolder = ViewHolder(v, layoutRes)
+    final override fun getViewHolder(v: View): ViewHolder =
+            ViewHolder(v, layoutRes)
 
     override fun failedToRecycle(holder: ViewHolder): Boolean {
         KL.e { "Failed to recycle" }
@@ -85,38 +97,59 @@ abstract class BindingItem<Binding : ViewDataBinding>(open val data: Any?) :
     override fun hashCode(): Int = data.hashCode()
 
     class ViewHolder(itemView: View, internal val layoutRes: Int) :
-        RecyclerView.ViewHolder(itemView)
+            RecyclerView.ViewHolder(itemView) {
+
+        /**
+         * Retrieves a binding.
+         *
+         * It is assumed that the binding is set prior to this call,
+         * and that its type matches the supplied generic.
+         */
+        fun <T> getBinding(): T = getBinding(itemView)
+    }
+
+    companion object {
+        fun setBinding(view: View, binding: Any) {
+            view.setTag(R.id.kau_view_binding_model, binding)
+        }
+
+        @Suppress("UNCHECKED_CAST")
+        fun <T> getBinding(view: View): T = view.getTag(R.id.kau_view_binding_model) as T
+    }
 }
 
-interface BindingLayout<Binding : ViewDataBinding> {
-    val layoutRes: Int
-}
-
-abstract class BindingClickEventHook<Binding : ViewDataBinding, Item : BindingItem<Binding>>(val identifier: BindingLayout<Binding>) :
-    ClickEventHook<Item>() {
+abstract class BindingClickEventHook<Binding : ViewBinding, Item : BindingItem<Binding>>(val identifier: BindingLayout<Binding>) :
+        ClickEventHook<Item>() {
 
     private fun RecyclerView.ViewHolder.binding(): Binding? {
         val holder = this as? BindingItem.ViewHolder ?: return null
         if (holder.layoutRes != identifier.layoutRes) {
             return null
         }
-        return DataBindingUtil.getBinding(itemView)
+        return getBinding()
     }
 
-    final override fun onBind(viewHolder: RecyclerView.ViewHolder): View? =
-        viewHolder.binding()?.onBind(viewHolder) ?: super.onBind(viewHolder)
+    final override fun onBind(viewHolder: RecyclerView.ViewHolder): View? {
+        val binding = viewHolder.binding() ?: return super.onBind(viewHolder)
+        val view = binding.onBind(viewHolder) ?: return super.onBind(viewHolder)
+        BindingItem.setBinding(view, binding)
+        return view
+    }
 
     open fun Binding.onBind(viewHolder: RecyclerView.ViewHolder): View? = super.onBind(viewHolder)
 
-    final override fun onBindMany(viewHolder: RecyclerView.ViewHolder): List<View>? =
-        viewHolder.binding()?.onBindMany(viewHolder) ?: super.onBindMany(viewHolder)
+    final override fun onBindMany(viewHolder: RecyclerView.ViewHolder): List<View>? {
+        val binding = viewHolder.binding() ?: return super.onBindMany(viewHolder)
+        val views = binding.onBindMany(viewHolder) ?: return super.onBindMany(viewHolder)
+        views.forEach { BindingItem.setBinding(it, binding) }
+        return views
+    }
 
     open fun Binding.onBindMany(viewHolder: RecyclerView.ViewHolder): List<View>? =
-        super.onBindMany(viewHolder)
+            super.onBindMany(viewHolder)
 
     final override fun onClick(v: View, position: Int, fastAdapter: FastAdapter<Item>, item: Item) {
-        val binding: Binding = DataBindingUtil.findBinding(v) ?: return
-        binding.onClick(v, position, fastAdapter, item)
+        BindingItem.getBinding<Binding>(v).onClick(v, position, fastAdapter, item)
     }
 
     abstract fun Binding.onClick(v: View, position: Int, fastAdapter: FastAdapter<Item>, item: Item)